Adding @dust-tt/types empty package
Remove readme file

Clean up

adjust name + tsconfig moduleResolution

Moving KeyType for POC

Add dts-cli

Fix front end github workflow

Adjust Docker build information for front

Adjust Dockerfile location

Testing new settings

npm install types

npm install

Fix types copy

Start command

Clean up

Fix format

Remove demo page

Cleaner build method

Deploy front edge adjusted

Fix _DOCKERFILE_PATH

Displaying initial pwd

print args

Fix deploy

Testing front and connectors deployment on edges

Add dts-cli to front/package-lock.json dependencies

Fix types/package-lock.json

Revert to connectors/package*.json from main

Harmonize Dockerfile and package-*.json

Build and lint types in build-and-lint-connectors

Adjust github workflow deploy-connectors-edge

Fix deploy-connectors-edge

Testing docker build for alerting/temporal without the deploy steps

Add build only Github workflow for alerting/temporal

Trigger workflow from branch

Adjust docker file path for alerting/temporal

Adjust all workflows

Fix connectors deployment

Add PostContentFragment io-ts

Format types

npm format

front> npm run format

Clean up after self review

types> add npm run lint to the CI build and lint workflow

npm run format

Downgrading typescript version for types/

fix front/package conflict
