+++ .github/workflows/build-and-lint-front.yml
@@ -16,5 +16,7 @@ jobs:
           node-version: 18.15.0
           cache: "npm"
           cache-dependency-path: ./front/package-lock.json
+      - working-directory: types
+        run: npm install && npm run lint &&  npm run build
       - working-directory: front
         run: npm install && npm run tsc && npm run lint && npm run format:check
