+++ .github/workflows/deploy-alerting-temporal.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Build the image on Cloud Build
         run: |
           chmod +x ./k8s/cloud-build.sh
-          ./k8s/cloud-build.sh alerting-temporal alerting/temporal
+          ./k8s/cloud-build.sh alerting-temporal ./Dockerfile ./alerting/temporal/
 
       - name: Deploy the image on Kubernetes
         run: |
