+++ .github/workflows/deploy-blog.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Build the image on Cloud Build
         run: |
           chmod +x ./k8s/cloud-build.sh
-          ./k8s/cloud-build.sh blog
+          ./k8s/cloud-build.sh blog ./Dockerfile ./blog/
 
       - name: Deploy the image on Kubernetes
         run: |
