+++ .github/workflows/deploy-connectors.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Build the image on Cloud Build
         run: |
           chmod +x ./k8s/cloud-build.sh
-          ./k8s/cloud-build.sh connectors
+          ./k8s/cloud-build.sh connectors ./connectors/Dockerfile ./
 
       - name: Deploy the image on Kubernetes
         run: |
