+++ .github/workflows/deploy-core.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Build the image on Cloud Build
         run: |
           chmod +x ./k8s/cloud-build.sh
-          ./k8s/cloud-build.sh core
+          ./k8s/cloud-build.sh core ./Dockerfile ./core/
 
       - name: Deploy the image on Kubernetes
         run: |
