+++ .github/workflows/deploy-docs.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Build the image on Cloud Build
         run: |
           chmod +x ./k8s/cloud-build.sh
-          ./k8s/cloud-build.sh docs
+          ./k8s/cloud-build.sh docs ./Dockerfile ./docs/
 
       - name: Deploy the image on Kubernetes
         run: |
