+++ .github/workflows/deploy-front.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Build the image on Cloud Build
         run: |
           chmod +x ./k8s/cloud-build.sh
-          ./k8s/cloud-build.sh front
+          ./k8s/cloud-build.sh front ./front/Dockerfile ../
 
       - name: Deploy the image on Kubernetes
         run: |
