+++ connectors/package.json
@@ -13,6 +13,7 @@
     "initdb": "env $(cat .env.local) npx tsx src/admin/db.ts"
   },
   "dependencies": {
+    "@dust-tt/types": "file:../types",
     "@nangohq/node": "^0.21.4",
     "@notionhq/client": "^2.2.4",
     "@slack/web-api": "^6.8.1",
