+++ front/Dockerfile
@@ -1,11 +1,17 @@
 FROM node:18.15.0 as front
 
-WORKDIR /app
-
-COPY package*.json ./
+WORKDIR /types
+COPY /types/package*.json ./
+COPY /types/ .
 RUN npm ci
 
-COPY . .
+WORKDIR /app
+
+COPY /front/package*.json ./
+RUN npm ci
+
+COPY /front .
+
 # fake database URIs are needed because Sequelize will throw if the `url` parameter
 # is undefined, and `next build` imports the `models.ts` file while "Collecting page data"
 RUN FRONT_DATABASE_URI="sqlite:foo.sqlite" XP1_DATABASE_URI="sqlite:bar.sqlite" npm run build
