+++ front/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@dust-tt/sparkle": "^0.2.47",
+    "@dust-tt/types": "file:../types",
     "@emoji-mart/data": "^1.1.2",
     "@emoji-mart/react": "^1.1.1",
     "@headlessui/react": "^1.7.7",
