+++ front/pages/api/v1/w/[wId]/assistant/conversations/[cId]/content_fragments.ts
@@ -1,3 +1,4 @@
+import { PostContentFragmentRequestBodySchema } from "@dust-tt/types";
 import { isLeft } from "fp-ts/lib/Either";
 import * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
@@ -15,25 +16,6 @@ import { ContentFragmentType } from "@app/types/assistant/conversation";
 export type PostContentFragmentsResponseBody = {
   contentFragment: ContentFragmentType;
 };
-
-export const PostContentFragmentRequestBodySchema = t.type({
-  title: t.string,
-  content: t.string,
-  url: t.union([t.string, t.null]),
-  contentType: t.union([
-    t.literal("slack_thread_content"),
-    t.literal("file_attachment"),
-  ]),
-  context: t.union([
-    t.type({
-      profilePictureUrl: t.union([t.string, t.null]),
-      fullName: t.union([t.string, t.null]),
-      email: t.union([t.string, t.null]),
-      username: t.union([t.string, t.null]),
-    }),
-    t.null,
-  ]),
-});
 
 export type PostContentFragmentRequestBody = t.TypeOf<
   typeof PostContentFragmentRequestBodySchema
