+++ front/pages/api/w/[wId]/keys/[secret]/disable.ts
@@ -1,9 +1,9 @@
+import { KeyType } from "@dust-tt/types";
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { Authenticator, getSession } from "@app/lib/auth";
 import { Key } from "@app/lib/models";
 import { withLogging } from "@app/logger/withlogging";
-import { KeyType } from "@app/types/key";
 
 export type PostKeysResponseBody = {
   key: KeyType;
