+++ types/.gitattributes
@@ -1,0 +1,1 @@
+* text eol=lf
