+++ types/.prettierrc
@@ -1,0 +1,7 @@
+{
+  "singleQuote": false,
+  "trailingComma": "es5",
+  "requirePragma": false,
+  "arrowParens": "always",
+  "semi": true
+}
