+++ types/src/index.ts
@@ -1,0 +1,5 @@
+import { KeyType } from "./front/types/key";
+
+import { PostContentFragmentRequestBodySchema } from "./front/pages/api/v1/w/[wId]/assistant/conversations/[cId]/content_fragments.types";
+export { KeyType };
+export { PostContentFragmentRequestBodySchema };
